Select language from login and register
