+++ src/components/GlobalHeader/RightContent.js
@@ -1,10 +1,11 @@
 import React, { PureComponent } from 'react';
-import { FormattedMessage, formatMessage, setLocale, getLocale } from 'umi/locale';
+import { FormattedMessage, formatMessage } from 'umi/locale';
 import { Spin, Tag, Menu, Icon, Dropdown, Avatar, Tooltip } from 'antd';
 import moment from 'moment';
 import groupBy from 'lodash/groupBy';
 import NoticeIcon from '../NoticeIcon';
 import HeaderSearch from '../HeaderSearch';
+import SelectLang from '../SelectLang';
 import styles from './index.less';
 
 export default class GlobalHeaderRight extends PureComponent {
@@ -39,10 +40,6 @@ export default class GlobalHeaderRight extends PureComponent {
     return groupBy(newNotices, 'type');
   }
 
-  changLang = ({ key }) => {
-    setLocale(key);
-  };
-
   render() {
     const {
       currentUser,
@@ -71,13 +68,6 @@ export default class GlobalHeaderRight extends PureComponent {
           <Icon type="logout" />
           <FormattedMessage id="menu.account.logout" defaultMessage="logout" />
         </Menu.Item>
-      </Menu>
-    );
-    const langMenu = (
-      <Menu className={styles.menu} selectedKeys={[getLocale()]} onClick={this.changLang}>
-        <Menu.Item key="zh-CN">中文</Menu.Item>
-        <Menu.Item key="en-US">English</Menu.Item>
-        <Menu.Item key="pt-BR">Português</Menu.Item>
       </Menu>
     );
     const noticeData = this.getNoticeData();
@@ -158,11 +148,7 @@ export default class GlobalHeaderRight extends PureComponent {
         ) : (
           <Spin size="small" style={{ marginLeft: 8, marginRight: 8 }} />
         )}
-        <Dropdown overlay={langMenu}>
-          <span className={styles.action}>
-            <FormattedMessage id="navbar.lang" /> <Icon type="down" />
-          </span>
-        </Dropdown>
+        <SelectLang className={styles.action} />
       </div>
     );
   }
