+++ src/components/SelectLang/index.less
@@ -1,0 +1,14 @@
+@import '~antd/lib/style/themes/default.less';
+
+.menu {
+  :global(.anticon) {
+    margin-right: 8px;
+  }
+  :global(.ant-dropdown-menu-item) {
+    width: 160px;
+  }
+}
+
+.dropdown {
+  cursor: pointer;
+}
