+++ src/layouts/UserLayout.less
@@ -8,6 +8,16 @@
   background: @layout-body-background;
 }
 
+.lang {
+  text-align: right;
+  width: 100%;
+  height: 40px;
+  line-height: 44px;
+  :global(.ant-dropdown-trigger) {
+    margin-right: 8px;
+  }
+}
+
 .content {
   padding: 32px 0;
   flex: 1;
@@ -22,7 +32,7 @@
   }
 
   .content {
-    padding: 112px 0 24px 0;
+    padding: 72px 0 24px 0;
   }
 }
